@@ -11,59 +11,63 @@
     "post-update": "echo \"codesandbox preview only, need an update\" && yarn upgrade --latest"
   },
   "dependencies": {
-    "@babel/core": "^7.12.10",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "autoprefixer": "^10.1.0",
-    "clean-css": "^4.2.3",
-    "axios": "^0.21.1",
-    "clsx": "^1.1.1",
-    "firebase": "^7.24.0",
-    "firebase-admin": "^9.5.0",
-    "formik": "^2.2.6",
-    "formik-material-ui": "^3.0.1",
-    "next": "^10.0.8",
-    "next-seo": "^4.17.0",
-    "postcss": "^8.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "@babel/core": "^7.20.7",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/material": "^5.11.2",
+    "@mui/styles": "^5.11.2",
+    "@mui/icons-material": "^5.11.0",
+    "autoprefixer": "^10.4.13",
+    "clean-css": "^5.3.1",
+    "axios": "^1.2.2",
+    "clsx": "^1.2.1",
+    "firebase": "^9.15.0",
+    "firebase-admin": "^11.4.1",
+    "formik": "^2.2.9",
+    "formik-mui": "^4.0.0",
+    "next": "^13.1.1",
+    "next-seo": "^5.15.0",
+    "postcss": "^8.4.20",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-is": "18.2.0",
     "react-lottie": "^1.2.3",
-    "react-pdf": "^5.2.0",
-    "sib-api-v3-typescript": "^2.0.2",
-    "styled-components": "^5.2.1"
+    "react-pdf": "^6.2.1",
+    "sib-api-v3-typescript": "^2.2.2",
+    "styled-components": "^5.3.6"
   },
   "devDependencies": {
-    "@fullhuman/postcss-purgecss": "^3.0.0",
+    "@fullhuman/postcss-purgecss": "^5.0.0",
     "@types/next": "^9.0.0",
-    "@types/node": "^14.14.14",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-lottie": "^1.2.5",
-    "@types/react-pdf": "^5.0.1",
-    "@types/styled-components": "^5.1.6",
-    "@typescript-eslint/eslint-plugin": "^4.14.0",
-    "@typescript-eslint/parser": "^4.14.0",
-    "adm-zip": "^0.5.1",
-    "eslint": "^7.18.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "firebase-tools": "^8.10.0",
-    "next-google-fonts": "^1.2.1",
-    "next-sitemap": "^1.3.16",
-    "node-fetch": "^2.6.0",
-    "p-queue": "^6.6.1",
-    "pdf-merger-js": "^3.0.5",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/react-lottie": "^1.2.6",
+    "@types/react-pdf": "^6.2.0",
+    "@types/styled-components": "^5.1.26",
+    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/parser": "^5.47.1",
+    "adm-zip": "^0.5.10",
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "firebase-tools": "^11.19.0",
+    "next-google-fonts": "^2.2.0",
+    "next-sitemap": "^3.1.43",
+    "node-fetch": "^3.3.0",
+    "p-queue": "^7.3.0",
+    "pdf-merger-js": "^4.2.0",
     "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "prettier": "^2.2.1",
-    "puppeteer": "^5.2.1",
-    "sharp": "^0.27.0",
-    "typescript": "^4.1.3",
+    "postcss-preset-env": "^7.8.3",
+    "prettier": "^2.8.1",
+    "puppeteer": "^19.4.1",
+    "sharp": "^0.31.3",
+    "typescript": "^4.9.4",
     "yarg": "^1.0.8",
-    "yargs": "^16.2.0"
+    "yargs": "^17.6.2"
   },
   "eslintConfig": {
     "extends": "react-app"
