@@ -1,6 +1,6 @@
 import React from 'react';
-import Typography from '@material-ui/core/Typography';
-import { Link } from '@material-ui/core';
+import Typography from '@mui/material/Typography';
+import { Link } from '@mui/material';
 import PageLayout from '../../src/PageLayout';
 import Config from '../../src/config/Config';
 
