@@ -1,8 +1,8 @@
-import { IconButton, IconButtonProps, styled } from '@material-ui/core';
+import { IconButton, IconButtonProps, styled } from '@mui/material';
 import React from 'react';
 import { logEvent } from '../../firebase';
 import Config from '../../config/Config';
-import Link from '@material-ui/core/Link';
+import Link from '@mui/material/Link';
 
 const StyledButton = styled(IconButton)(({ theme }) => ({
     alignSelf: 'center',
@@ -14,7 +14,12 @@ const StyledButton = styled(IconButton)(({ theme }) => ({
 
 const DonateButton: React.FC<IconButtonProps> = ({ ...props }) => (
     <Link href={Config.bmcUrl} target="_blank" rel="noopener" underline="none">
-        <StyledButton onClick={() => logEvent('bmc_click')} aria-label="buy me a sushi" {...props}>
+        <StyledButton
+            onClick={() => {
+                logEvent(/*'bmc_click'*/);
+            }}
+            aria-label="buy me a sushi"
+            {...props}>
             <span role={'img'} aria-label={'Sushi Emoji'}>
                 🍣
             </span>
