@@ -1,6 +1,6 @@
-import { createMuiTheme } from '@material-ui/core/styles';
+import { createTheme } from '@mui/material/styles';
 
-const appTheme = createMuiTheme({
+const appTheme = createTheme({
     typography: {
         allVariants: {
             fontFamily: 'Montserrat, sans-serif'
