@@ -1,11 +1,11 @@
-import firebase from 'firebase/app';
+//import firebase from 'firebase/app';
 import 'firebase/analytics';
 import 'firebase/performance';
 import 'firebase/storage';
-import Config from './config/Config';
+//import Config from './config/Config';
 
 function initFirebase(): void {
-    const enableFirebase = process.env.NEXT_PUBLIC_ENABLE_FIREBASE === 'true';
+    /*const enableFirebase = process.env.NEXT_PUBLIC_ENABLE_FIREBASE === 'true';
 
     if (enableFirebase && typeof window !== 'undefined' && !firebase.apps.length) {
         // Check that `window` is in scope for the analytics module!
@@ -15,12 +15,12 @@ function initFirebase(): void {
             firebase.performance();
         }
         firebase.storage();
-    }
+    }*/
 }
 
-function logEvent(eventName: string, eventParams?: Record<string, unknown>): void {
-    if (firebase.apps.length) firebase.analytics().logEvent(eventName, eventParams);
-    else console.log(`Event: (${eventName}) {${JSON.stringify(eventParams)}`);
+function logEvent(/*eventName: string, eventParams?: Record<string, unknown>*/): void {
+    //if (firebase.apps.length) firebase.analytics().logEvent(eventName, eventParams);
+    //else console.log(`Event: (${eventName}) {${JSON.stringify(eventParams)}`);
 }
 
 initFirebase();
